created a linear regression model
In this example, the data is loaded into a pandas DataFrame and then split into features and target. The train_test_split function is used to split the data into training and test sets. The linear regression model is trained using the fit method, and then the target values for the test data are predicted using the predict method. Finally, the mean squared error is calculated to evaluate the model's performance